Abstracted up inventory menus
